requirements.txt
flask
pillow
uvicorn
asgiref


